requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Pydantic
pydantic==2.10.3
pydantic-settings==2.6.1

# LangChain Core (Latest)
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13

# HuggingFace Integration (Latest)
langchain-huggingface==0.1.2
huggingface-hub==0.26.5
transformers==4.47.1

# Repository Management
gitpython==3.1.43

# Environment Management
python-dotenv==1.0.1

# Utilities
tiktoken==0.8.0

# Optional: For better performance
accelerate==1.2.1